Split electric collection endpoints into separate files
Move all "electric" collection-related API groups and their types out of api.ts into api/electric/collections as individual modules. This declutters the large api.ts file, groups related endpoints (attachments, channels, channelMembers, messages, messageReactions, invitations, pinnedMessages, notifications, organizations, organizationMembers, typingIndicators, directMessageParticipants, users) into dedicated files, and re-exports them from api/electric/collections/index.ts for backwards compatibility. The change was needed to organize collection endpoints, complete the attachment implementation, and make the API surface easier to maintain and navigate.
